Move burgr logic to a new class.
